pyproject.toml
[tool.poetry]
name = "red-tape-kit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "red_tape_kit" },
]

[tool.poetry.dependencies]
python = "^3.11"
fpdf2 = "^2.7.6"
python-docx = "^1.0.1"
beautifulsoup4 = "^4.12.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
factory_boy = "^3.3.0"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
